makes code more oriented
makes code more oriented
